+++ src/app/models/index.ts
@@ -41,7 +41,6 @@ export * from "./forms/action.form.model";
 export * from "./forms/contrast-test-parameters-action.form.model";
 export * from "./forms/stand-identification-action.form.model";
 export * from "./report/report.model";
-export * from "./report/report-body.model";
 export * from "./report/report-director.model";
-export * from "./report/report-body-contrast-test.model";
-export * from "./report/report-body-builder.model";
+export * from "./report/report-contrast-test.model";
+export * from "./report/report-builder.model";
