--- src/app/models/report/report-body.model.ts
@@ -1,3 +1,0 @@
-export class ReportBody {
-  public parts: string[] = [];
-}
