+++ src/app/models/report/report-builder.model.ts
@@ -1,0 +1,7 @@
+import { Report } from "./report.model";
+
+export class ReportBodyBuilder {
+  produce(): Report {
+    return new Report();
+  }
+}
