+++ src/app/models/steps/preparation.step.model.ts
@@ -5,7 +5,7 @@ import { StepBuilder } from "./step-builder.model";
 import { UserIdentificationAction } from "../actions/user-identification.action.model";
 import { StandIdentificationAction } from "../actions/stand-identification.action.model";
 import { StandIdentificationMinimalFieldsDecorator } from "../actions/meter-identification.action.model";
-import { ReportBodyBuilder } from "../report/report-body-builder.model";
+import { ReportBodyBuilder } from "../report/report-builder.model";
 
 export class PreparationStep extends StepBuilder {
